HV: - 64bit sizeof(SDK9_DRIVEINFO)[unpatched] was 4 bytes more than 32bit       sizeof(SDK9_DRIVEINFO), leading to a differently sized Mark5A/Mark5B       user directory, breaking cross-system portability.       Redefined SDK9_DRIVEINFO with compiler hints to force consistent       sizeof() across 32/64bit systems, added SDK9X_DRIVEINFO to support       any erroneously written user directories since the release of 64bit.     - Failure to force a specific user directory layout (for whatever       reason) left disk pack without user directory and VSN.       Albeit mostly a synthetical problem, operational impact is huge therefore       addressed and fixed.     - A failed "protect=off" would still allow the next command to execute       as if the pack had been unprotected.     - non-bank mode "status?" used a return type from SDK9 which was       incompatible with the return type in SDK8 for the same API call       XLRGetLength(). Both names resolve to the same underlying C/C++       data type so just changed to that to provide cross-SDK portability.

+++ evlbi5a/mk5command/status.cc
@@ -102,18 +102,27 @@ string status_fn(bool q, const vector<string>&, runtime& rte) {
         XLRCALL( ::XLRGetBankStatus(rte.xlrdev.sshandle(), BANK_B, &bs[1]) );
     } else if( bm==SS_BANKMODE_DISABLED ) {
         // In non-bank mode it's a tad more difficult
-        bool         status[ 2 ] = { true, true };
-        //S_DEVINFO    devInfo;
-        UINT64       length;
-        unsigned int nDisk[ 2 ]  = { 0   , 0    };
-
-        //XLRCALL( ::XLRGetDeviceInfo(GETSSHANDLE(rte), &devInfo) );
+        bool               status[ 2 ] = { true, true };
+        // SDK8 calls the return value of XLRGetLength() DWORDLONG, SDK9 calls it UINT64 
+        // and neiter bother to typedef an opaque type for an application
+        // writer that he/she can use to store the result of XLRGetLength()
+        // in. Inspection of the SDK8/SDK9 header files reveals that both
+        // are typedefs for 'unsigned long long'. So why don't we just
+        // forego the SDK's types and go for the underlying type?
+        // [because it's bad - but what can I do if Conduant decide to
+        //  change it yet again inna future 128 bit release??? (let's hope
+        //  we'll *never* get to see that day!)]
+        unsigned long long length; 
+        unsigned int       nDisk[ 2 ]  = { 0   , 0    };
 
         // How do we know that both packs are 'active'?
         // Maybe just loop over the number of buses?
         // one pack = 4 buses * 2 (master/slave)
         // so if we have >4 buses we're good?
         do_xlr_lock();
+
+        // We must get the recorded length in order to tell wether or not
+        // the disks are EMPTY (see below)
         length = ::XLRGetLength(GETSSHANDLE(rte));
 
         for(unsigned int nr = 0; nr<16; nr++) {
